+++ src/Features/Core/Portable/Diagnostics/DiagnosticAnalyzerExtensions.cs
@@ -8,8 +8,6 @@ using System.IO;
 using System.Linq;
 using System.Reflection;
 using Microsoft.CodeAnalysis.Diagnostics.Telemetry;
-using Microsoft.CodeAnalysis.Simplification;
-using Roslyn.Utilities;
 
 namespace Microsoft.CodeAnalysis.Diagnostics;
