+++ src/LanguageServer/Protocol/Features/Diagnostics/EngineV2/DiagnosticIncrementalAnalyzer_GetDiagnostics.cs
@@ -121,7 +121,7 @@ internal partial class DiagnosticAnalyzerService
                 if (s_projectToForceAnalysisData.TryGetValue(project.State, out var box) &&
                     analyzers.IsSubsetOf(box.Value.analyzers))
                 {
-                    var checksum = await StaticGetDiagnosticChecksumAsync(project, cancellationToken).ConfigureAwait(false);
+                    var checksum = await project.GetDiagnosticChecksumAsync(cancellationToken).ConfigureAwait(false);
                     if (box.Value.checksum == checksum)
                         return box.Value.diagnosticAnalysisResults;
                 }
