+++ src/LanguageServer/Protocol/Features/Diagnostics/EngineV2/DiagnosticIncrementalAnalyzer_IncrementalAnalyzer.cs
@@ -37,7 +37,7 @@ internal partial class DiagnosticAnalyzerService
         public async Task<ImmutableArray<DiagnosticData>> ForceAnalyzeProjectAsync(Project project, CancellationToken cancellationToken)
         {
             var projectState = project.State;
-            var checksum = await StaticGetDiagnosticChecksumAsync(project, cancellationToken).ConfigureAwait(false);
+            var checksum = await project.GetDiagnosticChecksumAsync(cancellationToken).ConfigureAwait(false);
 
             try
             {
