+++ src/LanguageServer/Protocol/Handler/Diagnostics/DiagnosticsPullCache.cs
@@ -15,16 +15,15 @@ namespace Microsoft.CodeAnalysis.LanguageServer.Handler.Diagnostics;
 internal abstract partial class AbstractPullDiagnosticHandler<TDiagnosticsParams, TReport, TReturn>
     where TDiagnosticsParams : IPartialResultParams<TReport>
 {
-    internal record struct DiagnosticsRequestState(
-        IDiagnosticAnalyzerService Service, Project Project, int GlobalStateVersion, RequestContext Context, IDiagnosticSource DiagnosticSource);
+    internal record struct DiagnosticsRequestState(Project Project, int GlobalStateVersion, RequestContext Context, IDiagnosticSource DiagnosticSource);
 
     /// <summary>
     /// Cache where we store the data produced by prior requests so that they can be returned if nothing of significance
     /// changed. The <see cref="VersionStamp"/> is produced by <see
     /// cref="Project.GetDependentVersionAsync(CancellationToken)"/> while the <see cref="Checksum"/> is produced by
-    /// <see cref="IDiagnosticAnalyzerService.GetDiagnosticChecksumAsync"/>.  The former is faster and works well for us
-    /// in the normal case.  The latter still allows us to reuse diagnostics when changes happen that update the version
-    /// stamp but not the content (for example, forking LSP text).
+    /// <see cref="CodeAnalysis.Diagnostics.Extensions.GetDiagnosticChecksumAsync"/>.  The former is faster and works
+    /// well for us in the normal case.  The latter still allows us to reuse diagnostics when changes happen that update
+    /// the version stamp but not the content (for example, forking LSP text).
     /// </summary>
     private sealed class DiagnosticsPullCache(string uniqueKey) : VersionedPullCache<(int globalStateVersion, VersionStamp? dependentVersion), (int globalStateVersion, Checksum dependentChecksum), DiagnosticsRequestState, ImmutableArray<DiagnosticData>>(uniqueKey)
     {
@@ -35,7 +34,7 @@ internal abstract partial class AbstractPullDiagnosticHandler<TDiagnosticsParams
 
         public override async Task<(int globalStateVersion, Checksum dependentChecksum)> ComputeExpensiveVersionAsync(DiagnosticsRequestState state, CancellationToken cancellationToken)
         {
-            return (state.GlobalStateVersion, await state.Service.GetDiagnosticChecksumAsync(state.Project, cancellationToken).ConfigureAwait(false));
+            return (state.GlobalStateVersion, await state.Project.GetDiagnosticChecksumAsync(cancellationToken).ConfigureAwait(false));
         }
 
         /// <inheritdoc cref="VersionedPullCache{TCheapVersion, TExpensiveVersion, TState, TComputedData}.ComputeDataAsync(TState, CancellationToken)"/>
